+++ cookbooks/icartapp/recipes/default.rb
@@ -1,0 +1,8 @@
+#
+# Cookbook Name:: icartapp
+# Recipe:: default
+#
+# Copyright 2014, YOUR_COMPANY_NAME
+#
+# All rights reserved - Do Not Redistribute
+#